1) remove classes LessonOnTimeTableDao/Service and DayOdWeekDao/Service
2) Add method from LessonOnTimeTableDao and DayOdWeekDao in TimeTableDao
class

3) Changed type of id variables into long

+++ src/main/java/ua/com/foxminded/rowmapper/ClassroomRowMapper.java
@@ -13,7 +13,7 @@ public class ClassroomRowMapper implements RowMapper<Classroom> {
     public Classroom mapRow(ResultSet rs, int rowNum) throws SQLException {
         Classroom classroom = new Classroom();
 
-        classroom.setId(rs.getInt(1));
+        classroom.setId(rs.getLong(1));
         classroom.setRoomNumber(rs.getInt(2));
         return classroom;
     }
