+++ src/main/java/ua/com/foxminded/rowmapper/GroupRowMapper.java
@@ -12,7 +12,7 @@ public class GroupRowMapper implements RowMapper<Group> {
     @Override
     public Group mapRow(ResultSet rs, int rowNum) throws SQLException {
         Group group = new Group();
-        group.setId(rs.getInt(1));
+        group.setId(rs.getLong(1));
         group.setGroupName(rs.getString(2));
         return group;
     }
